Add files via upload
in the second fundamental problem of today I had to reverse a number. The problem took 23 minutes to be solved but I solved it for a start calling the function with the number as a string between quotes. After that I had to learn how to make a number a string, so it too total 33 minutes; the approaches I learned today where number.toString() and String(number). Number being the variable for the function of course.